Stats: Reducing the code and colocating similar components.
