+++ config/client.json
@@ -30,6 +30,7 @@
 	"stats/plan-usage",
 	"stats/tier-upgrade-slider",
 	"stats/type-detection",
+	"stats/utm-module",
 	"wpcom_concierge_schedule_id",
 	"wpcom_signup_id",
 	"wpcom_signup_key",
